requirements.txt
geopandas==0.10.2
h2o==3.36.0.3
matplotlib==3.5.1
numba==0.55.1
numpy==1.21.5
GDAL==3.4.1
pandas==1.3.5
pyhdf==0.10.5
PyYAML==6.0
rasterio==1.2.10
scipy==1.7.3
Shapely==1.8.1.post1
tqdm==4.64.0
opencv-python==4.5.5.62
h5py==3.7.0
pyhdf==0.10.5
xlrd==2.0.1
openpyxl==3.0.9
tobler==0.9.0
rtree==1.0.0
astropy==4.3.1